Getting Program ready to merge with development
the Database function for changing fuel up data is not all that useful for now, deleting and reentering an entry can serve as a poor-man's update for now. Because of this I'd like to call this feature ready for the development branch so I can work on calculations which is the central focus of this project.