Trying and fixing the init calculator
Just a side note: I have a strong feeling that this implementation is very much a typical beginner solution. I'm kind of cooked right now so I don't think I can come up with a better solution for the time being.
I'll think about this for a little while and then come back to it.